+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to the "aiXcoder" extension will be documented in this file.
 
 ## [0.1.2] - %DATE%
 ### Added
+- Java: Automatically add missing class imports in completions.
 
 ### Changed
 - Use new telemetry API.